Update code to make it runnable
The codebase obtained from Jose Garmilla needed slight modifications to
make it runnable with the current LSST stack. These modifications
include changing the names of columns in catalogues to use underscores
instead of periods for separators, making some of the code compliant
with the LSST python formatting standard etc... The code is now able to
create and save a Star v/s Galaxy classifier object.